refactor: move authentication components to the "auth" module and update routes
